Moved PerFits library to header file
