--- tests/lib/CMakeLists.txt
@@ -1,3 +1,0 @@
-if ( ENABLE_FITS_TESTS )
-  add_subdirectory(PerFits)
-endif()
