+++ tests/utility/bitmap.hpp
